+++ Elements/pyGLV/GL/Scene.py
@@ -11,7 +11,8 @@ from typing import List, Dict
 
 from Elements.pyECSS.ECSSManager import ECSSManager
 #from Elements.pyGLV.GUI.Viewer import SDL2Window, ImGUIDecorator
-from Elements.pyGLV.GUI.Viewer import SDL2Window, ImGUIecssDecorator, ImGUIDecorator
+from Elements.pyGLV.GUI.Viewer import SDL2Window
+from Elements.pyGLV.GUI.ImguiDecorator import ImGUIecssDecorator, ImGUIDecorator
 
 class Scene():
     """
